requirements.txt
Flask
Flask-Cors
Flask-PyMongo
python-dotenv
bcrypt
pyjwt
